Renamed existing modals for the Transaction
