+++ src/app/transactions/actions/transactions-table-delete-modal.js
