+++ src/app/transactions/actions/transactions-table-edit-modal.js
